Refactor QuantisedDiatomic module
