+++ Project.toml
@@ -13,8 +13,8 @@ DiffEqCallbacks = "459566f4-90b8-5000-8ac3-15dfb0a30def"
 Distributions = "31c24e10-a181-5473-b8eb-7969acd0382f"
 FastBroadcast = "7034ab61-46d4-4ed7-9d0f-46aef9175898"
 HDF5 = "f67ccb44-e63f-5c2f-98bd-6dc0ccc4ba2f"
+Interpolations = "a98d9a8b-a2ab-59e6-89dd-64a1c18fca59"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
-LsqFit = "2fda8390-95c7-5789-9bda-21331edee243"
 MuladdMacro = "46d2c3a1-f734-5fdb-9937-b9b9aeba4221"
 NQCBase = "78c76ebc-5665-4934-b512-82d81b5cbfb7"
 NQCModels = "c814dc9f-a51f-4eaf-877f-82eda4edad48"
@@ -53,7 +53,7 @@ DiffEqCallbacks = "2.16"
 Distributions = "0.25"
 FastBroadcast = "0.1"
 HDF5 = "0.15, 0.16"
-LsqFit = "0.12"
+Interpolations = "0.13"
 MuladdMacro = "0.2"
 NQCBase = "0.2"
 NQCModels = "0.8"
