+++ test/InitialConditions/quantised_diatomic.jl
@@ -5,48 +5,52 @@ using LinearAlgebra: norm
 using Unitful
 using UnitfulAtomic
 
-atoms = Atoms([:N, :O])
-model = DiatomicHarmonic(r₀=2)
-sim = Simulation(atoms, model)
-surface = QuantisedDiatomic.SurfaceParameters(sim.atoms.masses, [1, 2], Matrix{Float64}(undef, 3, 0), 10.0, [0, 0, 1.0])
-
 @testset "separate/combine_slab_and_molecule" begin
-    atom_indices = [2, 3]
-    r0 = [
-        1 2 3 4 5;
-        6 7 8 9 10;
-        11 12 13 14 15
-    ]
-    molecule, slab = QuantisedDiatomic.separate_slab_and_molecule(atom_indices, r0)
-    @test molecule == [2 3; 7 8; 12 13]
-    @test slab == [1 4 5; 6 9 10; 11 14 15]
 
-    r = QuantisedDiatomic.combine_slab_and_molecule(atom_indices, molecule, slab)
-    @test r == r0
+    @testset "3D, with slab" begin
+        atom_indices = [2, 3]
+        r0 = [
+            1 2 3 4 5;
+            6 7 8 9 10;
+            11 12 13 14 15
+        ]
+        molecule, slab = QuantisedDiatomic.separate_slab_and_molecule(atom_indices, r0)
+        @test molecule == [2 3; 7 8; 12 13]
+        @test slab == [1 4 5; 6 9 10; 11 14 15]
 
-    atom_indices = [1, 2]
-    r0 = [1 2; 3 4; 5 6]
-    molecule, slab = QuantisedDiatomic.separate_slab_and_molecule(atom_indices, r0)
-    @test molecule == r0
+        r = QuantisedDiatomic.combine_slab_and_molecule(atom_indices, molecule, slab)
+        @test r == r0
+    end
 
-    r = QuantisedDiatomic.combine_slab_and_molecule(atom_indices, molecule, slab)
-    @test r == r0
-end
+    @testset "3D, no slab" begin
+        atom_indices = [1, 2]
+        r0 = [1 2; 3 4; 5 6]
+        molecule, slab = QuantisedDiatomic.separate_slab_and_molecule(atom_indices, r0)
+        @test molecule == r0
 
-@testset "calculate_diatomic_energy" begin
-    @test QuantisedDiatomic.calculate_diatomic_energy(model, 2.0, surface) ≈ 0.0 atol=1e-3
-    @test QuantisedDiatomic.calculate_diatomic_energy(model, 3.0, surface) ≈ 0.5
+        r = QuantisedDiatomic.combine_slab_and_molecule(atom_indices, molecule, slab)
+        @test r == r0
+    end
+
+    @testset "1D, no slab" begin
+        atom_indices = [1]
+        r0 = [1;;]
+        molecule, _ = QuantisedDiatomic.separate_slab_and_molecule(atom_indices, r0)
+        @test molecule == r0
+    end
 end
 
 @testset "calculate_force_constant" begin
     bond_lengths = 0.5:0.01:5.0
-    potential = QuantisedDiatomic.calculate_diatomic_energy.(model, bond_lengths, surface) # Calculate binding curve
+    model = NQCModels.DiatomicHarmonic(r₀=2.0)
+    environment = QuantisedDiatomic.EvaluationEnvironment([1, 2], (3, 2), zeros(3,0), 0.0, [0, 0, 1.0])
+    binding_curve = QuantisedDiatomic.calculate_binding_curve(bond_lengths, model, environment)
 
-    @test QuantisedDiatomic.calculate_force_constant(bond_lengths, potential)[1] ≈ 1 atol=0.1
+    @test QuantisedDiatomic.calculate_force_constant(binding_curve) ≈ 1 atol=0.1
 
-    model2 = NQCModels.DiatomicHarmonic(r₀=3.0)
-    potential = QuantisedDiatomic.calculate_diatomic_energy.(model2, bond_lengths, surface) # Calculate binding curve
-    @test QuantisedDiatomic.calculate_force_constant(bond_lengths, potential)[1] ≈ 1 atol=0.1
+    model = NQCModels.DiatomicHarmonic(r₀=3.0)
+    binding_curve = QuantisedDiatomic.calculate_binding_curve(bond_lengths, model, environment)
+    @test QuantisedDiatomic.calculate_force_constant(binding_curve) ≈ 1 atol=0.1
 end
 
 @testset "subtract_centre_of_mass!" begin
@@ -73,6 +77,10 @@ end
 end
 
 @testset "generate and check results" begin
+    atoms = Atoms([:N, :O])
+    model = DiatomicHarmonic(r₀=2)
+    sim = Simulation(atoms, model)
+
     numbers = [2, 5]
     configs = generate_configurations(sim, numbers[1], numbers[2]; samples=10, translational_energy=1u"eV")
     for config in configs
@@ -85,7 +93,7 @@ end
 @testset "position_above_surface!" begin
     r = zeros(3, 2)
     height = 10
-    QuantisedDiatomic.position_above_surface!(r, height, InfiniteCell())
+    QuantisedDiatomic.position_above_surface!(r, [0, 0, height], InfiniteCell())
     @test r == [0 0; 0 0; height height]
 end
 
@@ -96,6 +104,10 @@ end
 end
 
 @testset "apply_translational_impulse!" begin
+    atoms = Atoms([:N, :O])
+    model = DiatomicHarmonic(r₀=2)
+    sim = Simulation(atoms, model)
+
     masses = [100, 100]
     r = rand(3, 2) .+ [0 2; 0 2; 0 2]
     v = rand(3, 2) ./ 100
@@ -111,4 +123,101 @@ end
     νf, Jf = quantise_diatomic(sim, v, r)
     @test νi ≈ νf
     @test Ji ≈ Jf
+end
+
+@testset "find_total_energy" begin
+    environment = QuantisedDiatomic.EvaluationEnvironment([1, 2], (3, 2), zeros(3,0), 0.0, [0, 0, 1.0])
+    model = DiatomicHarmonic(r₀=2.0)
+    bond_lengths = 0.5:0.01:5.0
+    binding_curve = QuantisedDiatomic.calculate_binding_curve(bond_lengths, model, environment)
+    μ = 1.0
+    J = 2
+    ν = 1
+    V = QuantisedDiatomic.EffectivePotential(μ, J, binding_curve)
+    E, bounds = QuantisedDiatomic.find_total_energy(V, ν)
+end
+
+@testset "find_integral_bounds" begin
+    environment = QuantisedDiatomic.EvaluationEnvironment([1, 2], (3, 2), zeros(3,0), 0.0, [0, 0, 1.0])
+    model = DiatomicHarmonic(r₀=2)
+    bond_lengths = 0.5:0.01:5.0
+    binding_curve = QuantisedDiatomic.calculate_binding_curve(bond_lengths, model, environment)
+    total_energy = 1
+    J = 0
+    μ = 1.0
+    equilibrium_bond_length = 2
+    V = QuantisedDiatomic.EffectivePotential(μ, J, binding_curve)
+    r₁, r₂ = QuantisedDiatomic.find_integral_bounds(total_energy, V)
+    @test r₁ < equilibrium_bond_length
+    @test r₂ > equilibrium_bond_length
+end
+
+@testset "calculate_binding_curve" begin
+    model = Morse()
+    bond_lengths = 0.5:0.01:5.0
+    environment = QuantisedDiatomic.EvaluationEnvironment([1], (1, 1), zeros(1,0), 0.0, [0.0])
+    binding_curve = QuantisedDiatomic.calculate_binding_curve(bond_lengths, model, environment)
+    @test binding_curve.equilibrium_bond_length ≈ model.x₀
+end
+
+@testset "calculate_diatomic_energy" begin
+    @testset "3D, no slab" begin
+        model = DiatomicHarmonic(r₀=2)
+        environment = QuantisedDiatomic.EvaluationEnvironment([1, 2], (3, 2), zeros(3,0), 0.0, [0, 0, 1.0])
+        @test QuantisedDiatomic.calculate_diatomic_energy(2.0, model, environment) ≈ 0.0 atol=1e-3
+        @test QuantisedDiatomic.calculate_diatomic_energy(3.0, model, environment) ≈ 0.5
+    end
+
+    @testset "1D, no slab" begin
+        model = Harmonic(dofs=1)
+        bond_length = 1.5
+        environment = QuantisedDiatomic.EvaluationEnvironment([1], (1, 1), zeros(1,0), 0.0, [0.0])
+        @test QuantisedDiatomic.calculate_diatomic_energy(bond_length, model, environment) ≈ bond_length^2 / 2
+    end
+end
+
+@testset "assemble_evaluation_geometry" begin
+    bond_length = 4.3
+
+    @testset "3D, no slab" begin
+        environment = QuantisedDiatomic.EvaluationEnvironment([1, 2], (3, 4), rand(3,2), 2.0, [0, 0, 1.0])
+        r = QuantisedDiatomic.assemble_evaluation_geometry(bond_length, environment)
+        @test norm(r[:,1] .- r[:,2]) == 4.3
+    end
+
+    @testset "1D, no slab" begin
+        environment = QuantisedDiatomic.EvaluationEnvironment([1], (1, 1), zeros(1,0), 0.0, [0.0])
+        r = QuantisedDiatomic.assemble_evaluation_geometry(bond_length, environment)
+        @test r == [4.3;;]
+    end
+end
+
+@testset "build_molecule" begin
+    bond_length = 4.3
+
+    @testset "3D, no slab" begin
+        environment = QuantisedDiatomic.EvaluationEnvironment([1, 2], (3, 4), rand(3,2), 2.0, [0, 0, 1.0])
+        r = QuantisedDiatomic.build_molecule(bond_length, environment)
+        @test norm(r[:,1] .- r[:,2]) == 4.3
+    end
+
+    @testset "1D, no slab" begin
+        environment = QuantisedDiatomic.EvaluationEnvironment([1], (1, 1), zeros(0,0), 0.0, [0.0])
+        r = QuantisedDiatomic.build_molecule(bond_length, environment)
+        @test r == [4.3;;]
+    end
+end
+
+@testset "fit_binding_curve and plot_binding_curve" begin
+    model = Morse()
+    bond_lengths = 0.5:0.01:5.0
+    binding_curve = potential.(model, bond_lengths)
+    fit = QuantisedDiatomic.fit_binding_curve(bond_lengths, binding_curve)
+    QuantisedDiatomic.plot_binding_curve(bond_lengths, binding_curve, fit)
+end
+
+@testset "generate_configurations 1D" begin
+    sim = Simulation(Atoms(2000), Morse())
+    r, v = QuantisedDiatomic.generate_1D_vibrations(Morse(), 1000.0, 1; samples=10000)
+    QuantisedDiatomic.plot_distribution(r, v)
 end
